dpdk: move definition of main metadata structure to the end
Main metadata structure in the spec file may contain fields
which has structure type.
In such case the definition of those structures should be placed
before the definition of main metadata structure.
But currently the definitions of other structures are placed
after the definition of main metadata structure.
This commit moves the definition of main metadata structure
after the definitions of all other structures.